Add Area Defense calculations
Since Advantages can't give a Parry Bonus for a specific attack
this adds an Area Defense Technique.

This is based on Guns (Shotgun) as a Parry
so it gets halved, +3, + Combat Reflexes/ETS +1
and then further raised by Enhanced Parry (Area Defence)
or Enhanced Time Sense.

Then adds a Melee attack to the Gun
using the Area Defense skill as its base
but has to use it as an attack rather than a parry
to avoid halving it again.
The weapon then adds Gunslinger accuracy and RoF bonus.