Mike: Handle loaded weapon and spent ammo counting
Counting Gun loads in the sheet is usually more trouble than it's worth
since I can usually do it in my head on a per-encounter basis,
however gun weights include loaded ammo weight,
counting shots is importing for low-shots weapons and expensive bullets,
and counting the shots outside of the weapon double-counts the weight.

So to handle this the gun is converted into a container with the
right-click menu, and contents weight reduction is applied.

This isn't a perfect way to model it,
since in the absence of better ideas spent cartridges are assumed to
have negligible weight, and due to the weapon weight including the ammo
spent cartridges inside the gun weigh more in than carried.